src/tools/pdf_parser.py
"""
Lease Digitizer - PDF Parser Tool

A custom LangChain tool for parsing PDF documents and extracting text
with page-level metadata. Supports both text-based and scanned PDFs.

Key Features:
- High-quality text extraction with layout preservation
- Page-level text segmentation
- Metadata extraction (page count, author, dates)
- OCR fallback for scanned documents (planned)
- Memory-efficient streaming for large files
"""

from dataclasses import dataclass
from pathlib import Path
from typing import BinaryIO, Optional, Union

from langchain_core.tools import BaseTool
from pydantic import Field


@dataclass
class PageContent:
    """
    Extracted content from a single PDF page.
    
    Attributes:
        page_number: 1-indexed page number
        text: Extracted text content
        char_count: Number of characters
        has_tables: Whether tables were detected
        has_images: Whether images were detected
    """
    page_number: int
    text: str
    char_count: int
    has_tables: bool = False
    has_images: bool = False


@dataclass
class PDFMetadata:
    """
    Metadata extracted from a PDF document.
    
    Attributes:
        filename: Original filename
        page_count: Total number of pages
        author: Document author (if available)
        title: Document title (if available)
        creation_date: PDF creation date
        modification_date: PDF modification date
        file_size_bytes: File size in bytes
    """
    filename: str
    page_count: int
    author: Optional[str] = None
    title: Optional[str] = None
    creation_date: Optional[str] = None
    modification_date: Optional[str] = None
    file_size_bytes: Optional[int] = None


@dataclass
class ParsedDocument:
    """
    Complete parsed PDF document.
    
    Attributes:
        metadata: Document metadata
        pages: List of page contents
        full_text: Concatenated text from all pages
    """
    metadata: PDFMetadata
    pages: list[PageContent]
    full_text: str


class PDFParserTool(BaseTool):
    """
    LangChain tool for parsing PDF documents.
    
    Uses pypdf and optionally pdfplumber for enhanced extraction.
    Supports text-based PDFs with planned OCR support for scans.
    
    Example:
        >>> parser = PDFParserTool()
        >>> result = parser.parse("path/to/lease.pdf")
        >>> print(f"Pages: {result.metadata.page_count}")
    """
    
    name: str = "pdf_parser"
    description: str = (
        "Parse a PDF document and extract text content with metadata. "
        "Input should be a file path to a PDF document."
    )
    
    use_pdfplumber: bool = Field(
        default=True,
        description="Use pdfplumber for enhanced extraction"
    )
    
    def _run(self, file_path: str) -> str:
        """
        Parse a PDF and return extracted text.
        
        Args:
            file_path: Path to the PDF file
            
        Returns:
            Extracted text content
        """
        result = self.parse(file_path)
        return result.full_text
    
    async def _arun(self, file_path: str) -> str:
        """Async version of _run."""
        # TODO: Implement async parsing
        return self._run(file_path)
    
    def parse(self, file_path: Union[str, Path]) -> ParsedDocument:
        """
        Parse a PDF document and extract structured content.
        
        Args:
            file_path: Path to the PDF file
            
        Returns:
            ParsedDocument with metadata and page contents
            
        Raises:
            FileNotFoundError: If file doesn't exist
            ValueError: If file is not a valid PDF
        """
        path = Path(file_path)
        
        if not path.exists():
            raise FileNotFoundError(f"PDF file not found: {path}")
        
        if path.suffix.lower() != ".pdf":
            raise ValueError(f"File is not a PDF: {path}")
        
        # TODO: Implement PDF parsing
        # 1. Open PDF with pypdf/pdfplumber
        # 2. Extract metadata
        # 3. Extract text from each page
        # 4. Detect tables and images
        # 5. Build ParsedDocument
        
        raise NotImplementedError("PDF parsing not yet implemented")
    
    def parse_bytes(self, pdf_bytes: bytes, filename: str) -> ParsedDocument:
        """
        Parse a PDF from bytes.
        
        Args:
            pdf_bytes: PDF file content as bytes
            filename: Original filename for metadata
            
        Returns:
            ParsedDocument with metadata and page contents
        """
        # TODO: Implement bytes-based parsing
        raise NotImplementedError("Bytes parsing not yet implemented")
    
    def parse_stream(self, stream: BinaryIO, filename: str) -> ParsedDocument:
        """
        Parse a PDF from a file stream.
        
        Args:
            stream: File-like object containing PDF
            filename: Original filename for metadata
            
        Returns:
            ParsedDocument with metadata and page contents
        """
        # TODO: Implement stream-based parsing
        raise NotImplementedError("Stream parsing not yet implemented")
    
    def extract_text_with_layout(self, file_path: Union[str, Path]) -> str:
        """
        Extract text while preserving layout structure.
        
        Uses pdfplumber for layout-aware extraction.
        
        Args:
            file_path: Path to the PDF file
            
        Returns:
            Text with layout preserved
        """
        # TODO: Implement layout-preserving extraction
        raise NotImplementedError("Layout extraction not yet implemented")
    
    def extract_tables(self, file_path: Union[str, Path]) -> list[list[list[str]]]:
        """
        Extract tables from a PDF document.
        
        Args:
            file_path: Path to the PDF file
            
        Returns:
            List of tables, each as a 2D list of strings
        """
        # TODO: Implement table extraction
        raise NotImplementedError("Table extraction not yet implemented")


# TODO: Add OCR support using pytesseract or cloud OCR
# TODO: Add support for password-protected PDFs
# TODO: Add smart chunking for large documents
# TODO: Add caching for repeated parsing of same file
# TODO: Add parallel page processing for large PDFs
