requirements.txt
# Lease Digitizer - LangChain Multi-Agent System
# Production-grade document intelligence pipeline for commercial real estate leases

# Core LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# LLM Providers
openai>=1.10.0

# Document Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Data Validation & Schemas
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Store & Embeddings
chromadb>=0.4.22
tiktoken>=0.5.2

# Web Interface
streamlit>=1.29.0
streamlit-extras>=0.3.0

# Database & Persistence
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
structlog>=24.1.0
rich>=13.7.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
pre-commit>=3.6.0

# Type stubs
types-requests>=2.31.0
